refactor(shared): reorganize shared firebase and firestore modules (#892)
