+++ src/entities/card/api/subscription.spec.tsx
@@ -23,7 +23,10 @@ vi.mock("firebase/firestore", async (importOriginal) => {
     where: mocks.where,
   };
 });
-vi.mock("@/shared/firebase", () => ({ db: "db" }));
+vi.mock("@/shared/api", async (importOriginal) => ({
+  ...(await importOriginal<typeof import("@/shared/api")>()),
+  db: "db",
+}));
 
 import { subscribeCards } from "./firestore";
