+++ src/features/card-list/ui/CardListView.spec.tsx
@@ -13,7 +13,10 @@ import { describe, expect, it, vi } from "vitest";
 
 import { createCard } from "@/test/factories";
 
-vi.mock("@/shared/firebase", () => ({ auth: {} }));
+vi.mock("@/shared/api", async (importOriginal) => ({
+  ...(await importOriginal<typeof import("@/shared/api")>()),
+  auth: {},
+}));
 
 import { CardListView } from "./CardListView";
