+++ src/pages/card-form/ui/CardFormPage.spec.tsx
@@ -40,7 +40,10 @@ vi.mock("react-router-dom", () => ({
   useParams: () => mocks.params,
   useNavigate: () => mocks.navigate,
 }));
-vi.mock("@/shared/firebase", () => ({ auth: {} }));
+vi.mock("@/shared/api", async (importOriginal) => ({
+  ...(await importOriginal<typeof import("@/shared/api")>()),
+  auth: {},
+}));
 
 import { CardFormPage } from "./CardFormPage";
