+++ src/pages/deck-list/ui/DeckListPage.spec.tsx
@@ -81,7 +81,11 @@ vi.mock("@/features/deck-list", () => ({
   },
 }));
 vi.mock("react-router-dom", () => ({ useNavigate: () => mocks.navigate }));
-vi.mock("@/shared/firebase", () => ({ auth: {}, db: {} }));
+vi.mock("@/shared/api", async (importOriginal) => ({
+  ...(await importOriginal<typeof import("@/shared/api")>()),
+  auth: {},
+  db: {},
+}));
 
 import { DeckListPage } from "./DeckListPage";
