+++ src/pages/deck-swiper/ui/DeckSwiperPage.tsx
@@ -18,7 +18,7 @@ import {
   useStudyStore,
 } from "@/features/study";
 import { toggleShowHeader, toggleShowSwipeButtonList, usePreferences } from "@/entities/preferences";
-import { toRemoteById } from "@/shared/api";
+import { toRemoteById } from "@/shared/lib/remoteSnapshot";
 import { RouteFeedback } from "@/shared/ui/route-feedback";
 import { AppLayout } from "@/widgets/app-layout";
