+++ src/pages/deck-swiper/ui/DeckSwiperView.spec.tsx
@@ -2,7 +2,10 @@ import { render, screen } from "@testing-library/react";
 import "@testing-library/jest-dom/vitest";
 import { describe, expect, it, vi } from "vitest";
 
-vi.mock("@/shared/firebase", () => ({ auth: {} }));
+vi.mock("@/shared/api", async (importOriginal) => ({
+  ...(await importOriginal<typeof import("@/shared/api")>()),
+  auth: {},
+}));
 
 import { DeckSwiperView } from "./DeckSwiperView";
