+++ src/pages/settings/ui/SettingsPage.spec.tsx
@@ -15,7 +15,10 @@ const mocks = vi.hoisted(() => ({
   setDarkMode: vi.fn(),
 }));
 
-vi.mock("@/shared/firebase", () => ({ auth: {} }));
+vi.mock("@/shared/api", async (importOriginal) => ({
+  ...(await importOriginal<typeof import("@/shared/api")>()),
+  auth: {},
+}));
 vi.mock("@/entities/auth", () => ({ useAuthSession: () => mocks.authSession }));
 vi.mock("@/entities/preferences", () => ({
   usePreferences: () => mocks.preferences,
