+++ src/shared/api/firebase.ts
@@ -8,8 +8,8 @@ import { initializeApp } from "firebase/app";
 import { connectAuthEmulator, getAuth } from "firebase/auth";
 import { connectFirestoreEmulator, initializeFirestore, persistentLocalCache } from "firebase/firestore";
 
-const projectId = import.meta.env.VITE_PROJECT_ID;
-const apiKey = import.meta.env.VITE_WEB_API_KEY;
+const projectId = import.meta.env.VITE_PROJECT_ID || "demo-project";
+const apiKey = import.meta.env.VITE_WEB_API_KEY || "demo-key";
 
 const app = initializeApp({
   apiKey,
