+++ src/shared/api/index.ts
@@ -1,1 +1,7 @@
-export { toRemoteById } from "./remoteSnapshot";
+export { auth, db } from "./firebase";
+export {
+  firestoreTimestampDateSchema,
+  getTimestamp,
+  omitUndefined,
+  parseFirestoreDocument,
+} from "./firestoreDocument";
