--- src/shared/firestore/index.ts
@@ -1,6 +1,0 @@
-export {
-  firestoreTimestampDateSchema,
-  getTimestamp,
-  omitUndefined,
-  parseFirestoreDocument,
-} from "./firestoreDocument";
