+++ src/shared/lib/remoteSnapshot.ts
